Upgrades and fixes

+++ RoomHandler/RoomHandler/players/Stats.cs
@@ -3,7 +3,7 @@
     public short attack;
     public short upgradedAttack;
 
-    public float range;
+    public byte range;
     public byte upgradedRange;
     public byte fieldOfView;
 
@@ -14,7 +14,7 @@
 
     public byte level;
 
-    public Stats(short hp, short attack, short upgradedAttack, float range, byte fieldOfView, byte swordArmor, byte arrowArmor,
+    public Stats(short hp, short attack, short upgradedAttack, byte range, byte fieldOfView, byte swordArmor, byte arrowArmor,
         byte upgradedSwordArmor, byte upgradedArrowArmor, byte level) {
         this.hp = hp;
         this.attack = attack;
